requirements.txt
python-dateutil>=2.4.2
pyjwt>=1.1.0
requests>=2.4.0
six>=1.9.0
